==> 1.1_CREATE_CSV.py <==
import pandas as pd

list_of_labels = [  
                    'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama',
                    'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama',
                    'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama',
                    'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama',
                    'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama',
                    'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama',
                    'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama',
                    'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama',
                    'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama',
                    'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama', 'black_shama',
                    
                    'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 
                    'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike',
                    'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 
                    'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike',
                    'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 
                    'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike',
                    'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 
                    'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike',
                    'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 
                    'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike', 'blackish_cuckooshrike',
                   
                    'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite',
                    'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite',
                    'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite',
                    'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite',
                    'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite',
                    'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite',
                    'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite',
                    'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite',
                    'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite',
                    'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite', 'brahminy_kite',
                    
                    'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl',
                    'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl',
                    'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl',
                    'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl',
                    'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl',
                    'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl',
                    'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl',
                    'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl',
                    'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl',
                    'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl', 'brown_hawk_owl',
                    
                    'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet',
                    'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet',
                    'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet',
                    'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet',
                    'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet',
                    'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet',
                    'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet',
                    'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet',
                    'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet',
                    'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet', 'coppersmith_barbet',
                    
                    'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle',
                    'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle',
                    'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle',
                    'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle',
                    'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle',
                    'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle',
                    'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle',
                    'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle',
                    'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle',
                    'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle', 'crested_serpent_eagle',
                    
                    'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 
                    'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 
                    'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 
                    'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 
                    'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 
                    'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 
                    'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 
                    'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 
                    'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 
                    'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 'grey_nightjar', 
                    
                    'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 
                    'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 
                    'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 
                    'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 
                    'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 
                    'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 
                    'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 
                    'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 
                    'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 
                    'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 'orange_bellied_flowerpecker', 
                    
                    'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 
                    'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 
                    'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 
                    'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 
                    'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 
                    'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 
                    'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 
                    'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 
                    'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 
                    'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 'oriental_cuckoo', 
                    
                    'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 
                    'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 
                    'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 
                    'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 
                    'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 
                    'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 
                    'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 
                    'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 
                    'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 
                    'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker', 'white_bellied_woodpecker'
                ]
                 

list_of_audio_fns = [
                    'blackshama1.wav', 'blackshama2.wav', 'blackshama3.wav', 'blackshama4.wav', 'blackshama5.wav', 'blackshama6.wav', 'blackshama7.wav', 'blackshama8.wav', 'blackshama9.wav', 'blackshama10.wav', 
                    'blackshama11.wav', 'blackshama12.wav', 'blackshama13.wav', 'blackshama14.wav', 'blackshama15.wav', 'blackshama16.wav', 'blackshama17.wav', 'blackshama18.wav', 'blackshama19.wav', 'blackshama20.wav', 
                    'blackshama21.wav', 'blackshama22.wav', 'blackshama23.wav', 'blackshama24.wav', 'blackshama25.wav', 'blackshama26.wav', 'blackshama27.wav', 'blackshama28.wav', 'blackshama29.wav', 'blackshama30.wav', 
                    'blackshama31.wav', 'blackshama32.wav', 'blackshama33.wav', 'blackshama34.wav', 'blackshama35.wav', 'blackshama36.wav', 'blackshama37.wav', 'blackshama38.wav', 'blackshama39.wav', 'blackshama40.wav', 
                    'blackshama41.wav', 'blackshama42.wav', 'blackshama43.wav', 'blackshama44.wav', 'blackshama45.wav', 'blackshama46.wav', 'blackshama47.wav', 'blackshama48.wav', 'blackshama49.wav', 'blackshama50.wav', 
                    'blackshama51.wav', 'blackshama52.wav', 'blackshama53.wav', 'blackshama54.wav', 'blackshama55.wav', 'blackshama56.wav', 'blackshama57.wav', 'blackshama58.wav', 'blackshama59.wav', 'blackshama60.wav', 
                    'blackshama61.wav', 'blackshama62.wav', 'blackshama63.wav', 'blackshama64.wav', 'blackshama65.wav', 'blackshama66.wav', 'blackshama67.wav', 'blackshama68.wav', 'blackshama69.wav', 'blackshama70.wav', 
                    'blackshama71.wav', 'blackshama72.wav', 'blackshama73.wav', 'blackshama74.wav', 'blackshama75.wav', 'blackshama76.wav', 'blackshama77.wav', 'blackshama78.wav', 'blackshama79.wav', 'blackshama80.wav', 
                    'blackshama81.wav', 'blackshama82.wav', 'blackshama83.wav', 'blackshama84.wav', 'blackshama85.wav', 'blackshama86.wav', 'blackshama87.wav', 'blackshama88.wav', 'blackshama89.wav', 'blackshama90.wav', 
                    'blackshama91.wav', 'blackshama92.wav', 'blackshama93.wav', 'blackshama94.wav', 'blackshama95.wav', 'blackshama96.wav', 'blackshama97.wav', 'blackshama98.wav', 'blackshama99.wav', 'blackshama100.wav',
                    
                    'cuckooshrike1.wav',  'cuckooshrike2.wav',  'cuckooshrike3.wav',  'cuckooshrike4.wav',  'cuckooshrike5.wav',  'cuckooshrike6.wav',  'cuckooshrike7.wav',  'cuckooshrike8.wav',  'cuckooshrike9.wav',  'cuckooshrike10.wav', 
                    'cuckooshrike11.wav',  'cuckooshrike12.wav',  'cuckooshrike13.wav',  'cuckooshrike14.wav',  'cuckooshrike15.wav',  'cuckooshrike16.wav',  'cuckooshrike17.wav',  'cuckooshrike18.wav',  'cuckooshrike19.wav',  'cuckooshrike20.wav', 
                    'cuckooshrike21.wav',  'cuckooshrike22.wav',  'cuckooshrike23.wav',  'cuckooshrike24.wav',  'cuckooshrike25.wav',  'cuckooshrike26.wav',  'cuckooshrike27.wav',  'cuckooshrike28.wav',  'cuckooshrike29.wav',  'cuckooshrike30.wav', 
                    'cuckooshrike31.wav',  'cuckooshrike32.wav',  'cuckooshrike33.wav',  'cuckooshrike34.wav',  'cuckooshrike35.wav',  'cuckooshrike36.wav',  'cuckooshrike37.wav',  'cuckooshrike38.wav',  'cuckooshrike39.wav',  'cuckooshrike40.wav', 
                    'cuckooshrike41.wav',  'cuckooshrike42.wav',  'cuckooshrike43.wav',  'cuckooshrike44.wav',  'cuckooshrike45.wav',  'cuckooshrike46.wav',  'cuckooshrike47.wav',  'cuckooshrike48.wav',  'cuckooshrike49.wav',  'cuckooshrike50.wav', 
                    'cuckooshrike51.wav',  'cuckooshrike52.wav',  'cuckooshrike53.wav',  'cuckooshrike54.wav',  'cuckooshrike55.wav',  'cuckooshrike56.wav',  'cuckooshrike57.wav',  'cuckooshrike58.wav',  'cuckooshrike59.wav',  'cuckooshrike60.wav', 
                    'cuckooshrike61.wav',  'cuckooshrike62.wav',  'cuckooshrike63.wav',  'cuckooshrike64.wav',  'cuckooshrike65.wav',  'cuckooshrike66.wav',  'cuckooshrike67.wav',  'cuckooshrike68.wav',  'cuckooshrike69.wav',  'cuckooshrike70.wav', 
                    'cuckooshrike71.wav',  'cuckooshrike72.wav',  'cuckooshrike73.wav',  'cuckooshrike74.wav',  'cuckooshrike75.wav',  'cuckooshrike76.wav',  'cuckooshrike77.wav',  'cuckooshrike78.wav',  'cuckooshrike79.wav',  'cuckooshrike80.wav', 
                    'cuckooshrike81.wav',  'cuckooshrike82.wav',  'cuckooshrike83.wav',  'cuckooshrike84.wav',  'cuckooshrike85.wav',  'cuckooshrike86.wav',  'cuckooshrike87.wav',  'cuckooshrike88.wav',  'cuckooshrike89.wav',  'cuckooshrike90.wav', 
                    'cuckooshrike91.wav',  'cuckooshrike92.wav',  'cuckooshrike93.wav',  'cuckooshrike94.wav',  'cuckooshrike95.wav',  'cuckooshrike96.wav',  'cuckooshrike97.wav',  'cuckooshrike98.wav',  'cuckooshrike99.wav',  'cuckooshrike100.wav', 
                    
                    'brahminykite1.wav', 'brahminykite2.wav', 'brahminykite3.wav', 'brahminykite4.wav', 'brahminykite5.wav', 'brahminykite6.wav', 'brahminykite7.wav', 'brahminykite8.wav', 'brahminykite9.wav', 'brahminykite10.wav',
                    'brahminykite11.wav', 'brahminykite12.wav', 'brahminykite13.wav', 'brahminykite14.wav', 'brahminykite15.wav', 'brahminykite16.wav', 'brahminykite17.wav', 'brahminykite18.wav', 'brahminykite19.wav', 'brahminykite20.wav',
                    'brahminykite21.wav', 'brahminykite22.wav', 'brahminykite23.wav', 'brahminykite24.wav', 'brahminykite25.wav', 'brahminykite26.wav', 'brahminykite27.wav', 'brahminykite28.wav', 'brahminykite29.wav', 'brahminykite30.wav',
                    'brahminykite31.wav', 'brahminykite32.wav', 'brahminykite33.wav', 'brahminykite34.wav', 'brahminykite35.wav', 'brahminykite36.wav', 'brahminykite37.wav', 'brahminykite38.wav', 'brahminykite39.wav', 'brahminykite40.wav',
                    'brahminykite41.wav', 'brahminykite42.wav', 'brahminykite43.wav', 'brahminykite44.wav', 'brahminykite45.wav', 'brahminykite46.wav', 'brahminykite47.wav', 'brahminykite48.wav', 'brahminykite49.wav', 'brahminykite50.wav',
                    'brahminykite51.wav', 'brahminykite52.wav', 'brahminykite53.wav', 'brahminykite54.wav', 'brahminykite55.wav', 'brahminykite56.wav', 'brahminykite57.wav', 'brahminykite58.wav', 'brahminykite59.wav', 'brahminykite60.wav',
                    'brahminykite61.wav', 'brahminykite62.wav', 'brahminykite63.wav', 'brahminykite64.wav', 'brahminykite65.wav', 'brahminykite66.wav', 'brahminykite67.wav', 'brahminykite68.wav', 'brahminykite69.wav', 'brahminykite70.wav',
                    'brahminykite71.wav', 'brahminykite72.wav', 'brahminykite73.wav', 'brahminykite74.wav', 'brahminykite75.wav', 'brahminykite76.wav', 'brahminykite77.wav', 'brahminykite78.wav', 'brahminykite79.wav', 'brahminykite80.wav',
                    'brahminykite81.wav', 'brahminykite82.wav', 'brahminykite83.wav', 'brahminykite84.wav', 'brahminykite85.wav', 'brahminykite86.wav', 'brahminykite87.wav', 'brahminykite88.wav', 'brahminykite89.wav', 'brahminykite90.wav',
                    'brahminykite91.wav', 'brahminykite92.wav', 'brahminykite93.wav', 'brahminykite94.wav', 'brahminykite95.wav', 'brahminykite96.wav', 'brahminykite97.wav', 'brahminykite98.wav', 'brahminykite99.wav', 'brahminykite100.wav',
                    
                    'brownhawkowl1.wav', 'brownhawkowl2.wav', 'brownhawkowl3.wav', 'brownhawkowl4.wav', 'brownhawkowl5.wav', 'brownhawkowl6.wav', 'brownhawkowl7.wav', 'brownhawkowl8.wav', 'brownhawkowl9.wav', 'brownhawkowl10.wav',
                    'brownhawkowl11.wav', 'brownhawkowl12.wav', 'brownhawkowl13.wav', 'brownhawkowl14.wav', 'brownhawkowl15.wav', 'brownhawkowl16.wav', 'brownhawkowl17.wav', 'brownhawkowl18.wav', 'brownhawkowl19.wav', 'brownhawkowl20.wav',
                    'brownhawkowl21.wav', 'brownhawkowl22.wav', 'brownhawkowl23.wav', 'brownhawkowl24.wav', 'brownhawkowl25.wav', 'brownhawkowl26.wav', 'brownhawkowl27.wav', 'brownhawkowl28.wav', 'brownhawkowl29.wav', 'brownhawkowl30.wav',
                    'brownhawkowl31.wav', 'brownhawkowl32.wav', 'brownhawkowl33.wav', 'brownhawkowl34.wav', 'brownhawkowl35.wav', 'brownhawkowl36.wav', 'brownhawkowl37.wav', 'brownhawkowl38.wav', 'brownhawkowl39.wav', 'brownhawkowl40.wav',
                    'brownhawkowl41.wav', 'brownhawkowl42.wav', 'brownhawkowl43.wav', 'brownhawkowl44.wav', 'brownhawkowl45.wav', 'brownhawkowl46.wav', 'brownhawkowl47.wav', 'brownhawkowl48.wav', 'brownhawkowl49.wav', 'brownhawkowl50.wav',
                    'brownhawkowl51.wav', 'brownhawkowl52.wav', 'brownhawkowl53.wav', 'brownhawkowl54.wav', 'brownhawkowl55.wav', 'brownhawkowl56.wav', 'brownhawkowl57.wav', 'brownhawkowl58.wav', 'brownhawkowl59.wav', 'brownhawkowl60.wav',
                    'brownhawkowl61.wav', 'brownhawkowl62.wav', 'brownhawkowl63.wav', 'brownhawkowl64.wav', 'brownhawkowl65.wav', 'brownhawkowl66.wav', 'brownhawkowl67.wav', 'brownhawkowl68.wav', 'brownhawkowl69.wav', 'brownhawkowl70.wav',
                    'brownhawkowl71.wav', 'brownhawkowl72.wav', 'brownhawkowl73.wav', 'brownhawkowl74.wav', 'brownhawkowl75.wav', 'brownhawkowl76.wav', 'brownhawkowl77.wav', 'brownhawkowl78.wav', 'brownhawkowl79.wav', 'brownhawkowl80.wav',
                    'brownhawkowl81.wav', 'brownhawkowl82.wav', 'brownhawkowl83.wav', 'brownhawkowl84.wav', 'brownhawkowl85.wav', 'brownhawkowl86.wav', 'brownhawkowl87.wav', 'brownhawkowl88.wav', 'brownhawkowl89.wav', 'brownhawkowl90.wav',
                    'brownhawkowl91.wav', 'brownhawkowl92.wav', 'brownhawkowl93.wav', 'brownhawkowl94.wav', 'brownhawkowl95.wav', 'brownhawkowl96.wav', 'brownhawkowl97.wav', 'brownhawkowl98.wav', 'brownhawkowl99.wav', 'brownhawkowl100.wav',
                    
                    'barbet1.wav', 'barbet2.wav', 'barbet3.wav', 'barbet4.wav', 'barbet5.wav', 'barbet6.wav', 'barbet7.wav', 'barbet8.wav', 'barbet9.wav', 'barbet10.wav', 
                    'barbet11.wav', 'barbet12.wav', 'barbet13.wav', 'barbet14.wav', 'barbet15.wav', 'barbet16.wav', 'barbet17.wav', 'barbet18.wav', 'barbet19.wav', 'barbet20.wav', 
                    'barbet21.wav', 'barbet22.wav', 'barbet23.wav', 'barbet24.wav', 'barbet25.wav', 'barbet26.wav', 'barbet27.wav', 'barbet28.wav', 'barbet29.wav', 'barbet30.wav', 
                    'barbet31.wav', 'barbet32.wav', 'barbet33.wav', 'barbet34.wav', 'barbet35.wav', 'barbet36.wav', 'barbet37.wav', 'barbet38.wav', 'barbet39.wav', 'barbet40.wav', 
                    'barbet41.wav', 'barbet42.wav', 'barbet43.wav', 'barbet44.wav', 'barbet45.wav', 'barbet46.wav', 'barbet47.wav', 'barbet48.wav', 'barbet49.wav', 'barbet50.wav', 
                    'barbet51.wav', 'barbet52.wav', 'barbet53.wav', 'barbet54.wav', 'barbet55.wav', 'barbet56.wav', 'barbet57.wav', 'barbet58.wav', 'barbet59.wav', 'barbet60.wav', 
                    'barbet61.wav', 'barbet62.wav', 'barbet63.wav', 'barbet64.wav', 'barbet65.wav', 'barbet66.wav', 'barbet67.wav', 'barbet68.wav', 'barbet69.wav', 'barbet70.wav', 
                    'barbet71.wav', 'barbet72.wav', 'barbet73.wav', 'barbet74.wav', 'barbet75.wav', 'barbet76.wav', 'barbet77.wav', 'barbet78.wav', 'barbet79.wav', 'barbet80.wav', 
                    'barbet81.wav', 'barbet82.wav', 'barbet83.wav', 'barbet84.wav', 'barbet85.wav', 'barbet86.wav', 'barbet87.wav', 'barbet88.wav', 'barbet89.wav', 'barbet90.wav', 
                    'barbet91.wav', 'barbet92.wav', 'barbet93.wav', 'barbet94.wav', 'barbet95.wav', 'barbet96.wav', 'barbet97.wav', 'barbet98.wav', 'barbet99.wav', 'barbet100.wav',
                    
                    'cresentserpent1.wav', 'cresentserpent2.wav','cresentserpent3.wav', 'cresentserpent4.wav', 'cresentserpent5.wav', 'cresentserpent6.wav', 'cresentserpent7.wav', 'cresentserpent8.wav', 'cresentserpent9.wav', 'cresentserpent10.wav',  
                    'cresentserpent11.wav', 'cresentserpent12.wav','cresentserpent13.wav', 'cresentserpent14.wav', 'cresentserpent15.wav', 'cresentserpent16.wav', 'cresentserpent17.wav', 'cresentserpent18.wav', 'cresentserpent19.wav', 'cresentserpent20.wav',  
                    'cresentserpent21.wav', 'cresentserpent22.wav','cresentserpent23.wav', 'cresentserpent24.wav', 'cresentserpent25.wav', 'cresentserpent26.wav', 'cresentserpent27.wav', 'cresentserpent28.wav', 'cresentserpent29.wav', 'cresentserpent30.wav',  
                    'cresentserpent31.wav', 'cresentserpent32.wav','cresentserpent33.wav', 'cresentserpent34.wav', 'cresentserpent35.wav', 'cresentserpent36.wav', 'cresentserpent37.wav', 'cresentserpent38.wav', 'cresentserpent39.wav', 'cresentserpent40.wav',  
                    'cresentserpent41.wav', 'cresentserpent42.wav','cresentserpent43.wav', 'cresentserpent44.wav', 'cresentserpent45.wav', 'cresentserpent46.wav', 'cresentserpent47.wav', 'cresentserpent48.wav', 'cresentserpent49.wav', 'cresentserpent50.wav',  
                    'cresentserpent51.wav', 'cresentserpent52.wav','cresentserpent53.wav', 'cresentserpent54.wav', 'cresentserpent55.wav', 'cresentserpent56.wav', 'cresentserpent57.wav', 'cresentserpent58.wav', 'cresentserpent59.wav', 'cresentserpent60.wav',  
                    'cresentserpent61.wav', 'cresentserpent62.wav','cresentserpent63.wav', 'cresentserpent64.wav', 'cresentserpent65.wav', 'cresentserpent66.wav', 'cresentserpent67.wav', 'cresentserpent68.wav', 'cresentserpent69.wav', 'cresentserpent70.wav',  
                    'cresentserpent71.wav', 'cresentserpent72.wav','cresentserpent73.wav', 'cresentserpent74.wav', 'cresentserpent75.wav', 'cresentserpent76.wav', 'cresentserpent77.wav', 'cresentserpent78.wav', 'cresentserpent79.wav', 'cresentserpent80.wav',  
                    'cresentserpent81.wav', 'cresentserpent82.wav','cresentserpent83.wav', 'cresentserpent84.wav', 'cresentserpent85.wav', 'cresentserpent86.wav', 'cresentserpent87.wav', 'cresentserpent88.wav', 'cresentserpent89.wav', 'cresentserpent90.wav',  
                    'cresentserpent91.wav', 'cresentserpent92.wav','cresentserpent93.wav', 'cresentserpent94.wav', 'cresentserpent95.wav', 'cresentserpent96.wav', 'cresentserpent97.wav', 'cresentserpent98.wav', 'cresentserpent99.wav', 'cresentserpent100.wav',  
                    
                    'greynightjar1.wav', 'greynightjar2.wav', 'greynightjar3.wav', 'greynightjar4.wav', 'greynightjar5.wav', 'greynightjar6.wav', 'greynightjar7.wav', 'greynightjar8.wav', 'greynightjar9.wav', 'greynightjar10.wav', 
                    'greynightjar11.wav', 'greynightjar12.wav', 'greynightjar13.wav', 'greynightjar14.wav', 'greynightjar15.wav', 'greynightjar16.wav', 'greynightjar17.wav', 'greynightjar18.wav', 'greynightjar19.wav', 'greynightjar20.wav', 
                    'greynightjar21.wav', 'greynightjar22.wav', 'greynightjar23.wav', 'greynightjar24.wav', 'greynightjar25.wav', 'greynightjar26.wav', 'greynightjar27.wav', 'greynightjar28.wav', 'greynightjar29.wav', 'greynightjar30.wav', 
                    'greynightjar31.wav', 'greynightjar32.wav', 'greynightjar33.wav', 'greynightjar34.wav', 'greynightjar35.wav', 'greynightjar36.wav', 'greynightjar37.wav', 'greynightjar38.wav', 'greynightjar39.wav', 'greynightjar40.wav', 
                    'greynightjar41.wav', 'greynightjar42.wav', 'greynightjar43.wav', 'greynightjar44.wav', 'greynightjar45.wav', 'greynightjar46.wav', 'greynightjar47.wav', 'greynightjar48.wav', 'greynightjar49.wav', 'greynightjar50.wav', 
                    'greynightjar51.wav', 'greynightjar52.wav', 'greynightjar53.wav', 'greynightjar54.wav', 'greynightjar55.wav', 'greynightjar56.wav', 'greynightjar57.wav', 'greynightjar58.wav', 'greynightjar59.wav', 'greynightjar60.wav', 
                    'greynightjar61.wav', 'greynightjar62.wav', 'greynightjar63.wav', 'greynightjar64.wav', 'greynightjar65.wav', 'greynightjar66.wav', 'greynightjar67.wav', 'greynightjar68.wav', 'greynightjar69.wav', 'greynightjar70.wav', 
                    'greynightjar71.wav', 'greynightjar72.wav', 'greynightjar73.wav', 'greynightjar74.wav', 'greynightjar75.wav', 'greynightjar76.wav', 'greynightjar77.wav', 'greynightjar78.wav', 'greynightjar79.wav', 'greynightjar80.wav', 
                    'greynightjar81.wav', 'greynightjar82.wav', 'greynightjar83.wav', 'greynightjar84.wav', 'greynightjar85.wav', 'greynightjar86.wav', 'greynightjar87.wav', 'greynightjar88.wav', 'greynightjar89.wav', 'greynightjar90.wav', 
                    'greynightjar91.wav', 'greynightjar92.wav', 'greynightjar93.wav', 'greynightjar94.wav', 'greynightjar95.wav', 'greynightjar96.wav', 'greynightjar97.wav', 'greynightjar98.wav', 'greynightjar99.wav', 'greynightjar100.wav', 
                    
                    'flowerpecker1.wav', 'flowerpecker2.wav', 'flowerpecker3.wav', 'flowerpecker4.wav', 'flowerpecker5.wav', 'flowerpecker6.wav', 'flowerpecker7.wav', 'flowerpecker8.wav', 'flowerpecker9.wav', 'flowerpecker10.wav', 
                    'flowerpecker11.wav', 'flowerpecker12.wav', 'flowerpecker13.wav', 'flowerpecker14.wav', 'flowerpecker15.wav', 'flowerpecker16.wav', 'flowerpecker17.wav', 'flowerpecker18.wav', 'flowerpecker19.wav', 'flowerpecker20.wav', 
                    'flowerpecker21.wav', 'flowerpecker22.wav', 'flowerpecker23.wav', 'flowerpecker24.wav', 'flowerpecker25.wav', 'flowerpecker26.wav', 'flowerpecker27.wav', 'flowerpecker28.wav', 'flowerpecker29.wav', 'flowerpecker30.wav', 
                    'flowerpecker31.wav', 'flowerpecker32.wav', 'flowerpecker33.wav', 'flowerpecker34.wav', 'flowerpecker35.wav', 'flowerpecker36.wav', 'flowerpecker37.wav', 'flowerpecker38.wav', 'flowerpecker39.wav', 'flowerpecker40.wav', 
                    'flowerpecker41.wav', 'flowerpecker42.wav', 'flowerpecker43.wav', 'flowerpecker44.wav', 'flowerpecker45.wav', 'flowerpecker46.wav', 'flowerpecker47.wav', 'flowerpecker48.wav', 'flowerpecker49.wav', 'flowerpecker50.wav', 
                    'flowerpecker51.wav', 'flowerpecker52.wav', 'flowerpecker53.wav', 'flowerpecker54.wav', 'flowerpecker55.wav', 'flowerpecker56.wav', 'flowerpecker57.wav', 'flowerpecker58.wav', 'flowerpecker59.wav', 'flowerpecker60.wav', 
                    'flowerpecker61.wav', 'flowerpecker62.wav', 'flowerpecker63.wav', 'flowerpecker64.wav', 'flowerpecker65.wav', 'flowerpecker66.wav', 'flowerpecker67.wav', 'flowerpecker68.wav', 'flowerpecker69.wav', 'flowerpecker70.wav', 
                    'flowerpecker71.wav', 'flowerpecker72.wav', 'flowerpecker73.wav', 'flowerpecker74.wav', 'flowerpecker75.wav', 'flowerpecker76.wav', 'flowerpecker77.wav', 'flowerpecker78.wav', 'flowerpecker79.wav', 'flowerpecker80.wav', 
                    'flowerpecker81.wav', 'flowerpecker82.wav', 'flowerpecker83.wav', 'flowerpecker84.wav', 'flowerpecker85.wav', 'flowerpecker86.wav', 'flowerpecker87.wav', 'flowerpecker88.wav', 'flowerpecker89.wav', 'flowerpecker90.wav', 
                    'flowerpecker91.wav', 'flowerpecker92.wav', 'flowerpecker93.wav', 'flowerpecker94.wav', 'flowerpecker95.wav', 'flowerpecker96.wav', 'flowerpecker97.wav', 'flowerpecker98.wav', 'flowerpecker99.wav', 'flowerpecker100.wav', 
                    
                    'orientalcuckoo1.wav', 'orientalcuckoo2.wav', 'orientalcuckoo3.wav', 'orientalcuckoo4.wav', 'orientalcuckoo5.wav', 'orientalcuckoo6.wav', 'orientalcuckoo7.wav', 'orientalcuckoo8.wav', 'orientalcuckoo9.wav', 'orientalcuckoo10.wav', 
                    'orientalcuckoo11.wav', 'orientalcuckoo12.wav', 'orientalcuckoo13.wav', 'orientalcuckoo14.wav', 'orientalcuckoo15.wav', 'orientalcuckoo16.wav', 'orientalcuckoo17.wav', 'orientalcuckoo18.wav', 'orientalcuckoo19.wav', 'orientalcuckoo20.wav', 
                    'orientalcuckoo21.wav', 'orientalcuckoo22.wav', 'orientalcuckoo23.wav', 'orientalcuckoo24.wav', 'orientalcuckoo25.wav', 'orientalcuckoo26.wav', 'orientalcuckoo27.wav', 'orientalcuckoo28.wav', 'orientalcuckoo29.wav', 'orientalcuckoo30.wav', 
                    'orientalcuckoo31.wav', 'orientalcuckoo32.wav', 'orientalcuckoo33.wav', 'orientalcuckoo34.wav', 'orientalcuckoo35.wav', 'orientalcuckoo36.wav', 'orientalcuckoo37.wav', 'orientalcuckoo38.wav', 'orientalcuckoo39.wav', 'orientalcuckoo40.wav', 
                    'orientalcuckoo41.wav', 'orientalcuckoo42.wav', 'orientalcuckoo43.wav', 'orientalcuckoo44.wav', 'orientalcuckoo45.wav', 'orientalcuckoo46.wav', 'orientalcuckoo47.wav', 'orientalcuckoo48.wav', 'orientalcuckoo49.wav', 'orientalcuckoo50.wav', 
                    'orientalcuckoo51.wav', 'orientalcuckoo52.wav', 'orientalcuckoo53.wav', 'orientalcuckoo54.wav', 'orientalcuckoo55.wav', 'orientalcuckoo56.wav', 'orientalcuckoo57.wav', 'orientalcuckoo58.wav', 'orientalcuckoo59.wav', 'orientalcuckoo60.wav', 
                    'orientalcuckoo61.wav', 'orientalcuckoo62.wav', 'orientalcuckoo63.wav', 'orientalcuckoo64.wav', 'orientalcuckoo65.wav', 'orientalcuckoo66.wav', 'orientalcuckoo67.wav', 'orientalcuckoo68.wav', 'orientalcuckoo69.wav', 'orientalcuckoo70.wav', 
                    'orientalcuckoo71.wav', 'orientalcuckoo72.wav', 'orientalcuckoo73.wav', 'orientalcuckoo74.wav', 'orientalcuckoo75.wav', 'orientalcuckoo76.wav', 'orientalcuckoo77.wav', 'orientalcuckoo78.wav', 'orientalcuckoo79.wav', 'orientalcuckoo80.wav', 
                    'orientalcuckoo81.wav', 'orientalcuckoo82.wav', 'orientalcuckoo83.wav', 'orientalcuckoo84.wav', 'orientalcuckoo85.wav', 'orientalcuckoo86.wav', 'orientalcuckoo87.wav', 'orientalcuckoo88.wav', 'orientalcuckoo89.wav', 'orientalcuckoo90.wav', 
                    'orientalcuckoo91.wav', 'orientalcuckoo92.wav', 'orientalcuckoo93.wav', 'orientalcuckoo94.wav', 'orientalcuckoo95.wav', 'orientalcuckoo96.wav', 'orientalcuckoo97.wav', 'orientalcuckoo98.wav', 'orientalcuckoo99.wav', 'orientalcuckoo100.wav',
                    
                    'woodpecker1.wav', 'woodpecker2.wav', 'woodpecker3.wav', 'woodpecker4.wav', 'woodpecker5.wav', 'woodpecker6.wav', 'woodpecker7.wav', 'woodpecker8.wav', 'woodpecker9.wav', 'woodpecker10.wav', 
                    'woodpecker11.wav', 'woodpecker12.wav', 'woodpecker13.wav', 'woodpecker14.wav', 'woodpecker15.wav', 'woodpecker16.wav', 'woodpecker17.wav', 'woodpecker18.wav', 'woodpecker19.wav', 'woodpecker20.wav', 
                    'woodpecker21.wav', 'woodpecker22.wav', 'woodpecker23.wav', 'woodpecker24.wav', 'woodpecker25.wav', 'woodpecker26.wav', 'woodpecker27.wav', 'woodpecker28.wav', 'woodpecker29.wav', 'woodpecker30.wav', 
                    'woodpecker31.wav', 'woodpecker32.wav', 'woodpecker33.wav', 'woodpecker34.wav', 'woodpecker35.wav', 'woodpecker36.wav', 'woodpecker37.wav', 'woodpecker38.wav', 'woodpecker39.wav', 'woodpecker40.wav', 
                    'woodpecker41.wav', 'woodpecker42.wav', 'woodpecker43.wav', 'woodpecker44.wav', 'woodpecker45.wav', 'woodpecker46.wav', 'woodpecker47.wav', 'woodpecker48.wav', 'woodpecker49.wav', 'woodpecker50.wav', 
                    'woodpecker51.wav', 'woodpecker52.wav', 'woodpecker53.wav', 'woodpecker54.wav', 'woodpecker55.wav', 'woodpecker56.wav', 'woodpecker57.wav', 'woodpecker58.wav', 'woodpecker59.wav', 'woodpecker60.wav', 
                    'woodpecker61.wav', 'woodpecker62.wav', 'woodpecker63.wav', 'woodpecker64.wav', 'woodpecker65.wav', 'woodpecker66.wav', 'woodpecker67.wav', 'woodpecker68.wav', 'woodpecker69.wav', 'woodpecker70.wav', 
                    'woodpecker71.wav', 'woodpecker72.wav', 'woodpecker73.wav', 'woodpecker74.wav', 'woodpecker75.wav', 'woodpecker76.wav', 'woodpecker77.wav', 'woodpecker78.wav', 'woodpecker79.wav', 'woodpecker80.wav', 
                    'woodpecker81.wav', 'woodpecker82.wav', 'woodpecker83.wav', 'woodpecker84.wav', 'woodpecker85.wav', 'woodpecker86.wav', 'woodpecker87.wav', 'woodpecker88.wav', 'woodpecker89.wav', 'woodpecker90.wav', 
                    'woodpecker91.wav', 'woodpecker92.wav', 'woodpecker93.wav', 'woodpecker94.wav', 'woodpecker95.wav', 'woodpecker96.wav', 'woodpecker97.wav', 'woodpecker98.wav', 'woodpecker99.wav', 'woodpecker100.wav', 
                    
                ]

X = []
print(len(list_of_labels), len(list_of_audio_fns))

for fn, label in zip(list_of_audio_fns, list_of_labels):
    X.append([fn, label])

df = pd.DataFrame(X)
df.columns = ['filename', 'label']
df.to_csv('BIRDS.csv', index = False)
